Replaced reservoir sampling with shuffling.
Unfortunately, the reservoir sampling implementation was bugged, and correcting it would make the algorithm `O(max(n, k))`.
This is due to not being able to advance the set iterator efficiently by a delta, which makes it so we need to call `next(it)` in a loop.

As for shuffling, we can't make it more efficient by only shuffling `k` elements, as there's no built-in function for that, and slicing the list in any way will affect the results.